add Bookmark.js and add const form with eventListener to index.js
