+++ js-structure/js-structure_quiz-app/index.js
@@ -1,4 +1,6 @@
 import Header from "./components/Header/Header.js";
+import Form from "./components/Form/Form.js";
+import Bookmark from "./components/Bookmark/Bookmark.js";
 
 // Store the cards in a global state
 const cards = [
@@ -26,7 +28,8 @@ const cards = [
 ];
 
 
-
+const form = Form()
+form.addEventListener("submit", handleFormSubmit)
 function handleFormSubmit(event) {
   event.preventDefault();
 
@@ -51,81 +54,8 @@ function handleFormSubmit(event) {
   event.target.elements.question.focus();
 }
 
-function Form() {
-  const form = document.createElement("form");
-  form.classList.add("form");
-  form.innerHTML = /* html */ `
-		<label for="question">Your question:</label>
-		<textarea
-			id="question"
-			rows="7"
-			name="question"
-			maxlength="150"
-			data-js="input-question"
-		></textarea>
-		<span class="form__character-count">
-			<span data-js="amount-left-question"></span> characters left
-		</span>
-		<label for="answer">Your answer:</label>
-		<textarea
-			id="answer"
-			rows="7"
-			name="answer"
-			maxlength="150"
-			data-js="input-answer"
-		></textarea>
-		<span class="form__character-count">
-			<span data-js="amount-left-answer"></span> characters left
-		</span>
-		<label for="tags">Tags:</label>
-		<input id="tags" type="text" name="tags" />
-		<button class="form__submit-button" type="submit">Submit</button>
-	`;
 
-  const formFields = form.querySelectorAll('[data-js*="input"]');
-  const counterOutputs = form.querySelectorAll("[data-js*=amount-left]");
 
-  formFields.forEach((formField, index) => {
-    counterOutputs.forEach((output) => {
-      output.innerText = formField.maxLength;
-    });
-    formField.addEventListener("input", () => {
-      const currentAmountLeft = formField.maxLength - formField.value.length;
-      counterOutputs[index].innerText = currentAmountLeft;
-    });
-  });
-
-  form.addEventListener("submit", handleFormSubmit);
-
-  return form;
-}
-
-function Bookmark(props) {
-  const bookmark = document.createElement("button");
-  bookmark.classList.add("bookmark");
-
-  if (props.active) {
-    bookmark.classList.add("bookmark--active");
-  }
-
-  bookmark.type = "button";
-  bookmark.setAttribute("aria-label", "bookmark");
-  bookmark.innerHTML = /* html */ `
-		<svg
-			class="bookmark__icon"
-			xmlns="http://www.w3.org/2000/svg"
-			viewbox="0 0 24 24"
-		>
-			<path
-				d="M17 3H7c-1.1 0-2 .9-2 2v16l7-3 7 3V5c0-1.1-.9-2-2-2z"
-			/>
-		</svg>
-	`;
-
-  bookmark.addEventListener("click", props.onClick);
-
-  return bookmark;
-}
 
 function Card(props) {
   const card = document.createElement("article");
